server/requirements.txt
flask
flask-cors
pillow
pyyaml
